split tip parser into several files
